moved pcb_cache to separate files and other comments addressed
